[packaging] Update patches for PulseAudio 6.0.
[packaging] Update patches for PulseAudio 6.0. Contributes to JB#26620

Some comments for new patches.

Patch 1012-bluez4-device-Don-t-acquire-transport-when-connectin.patch:

Previously when creating A2DP Bluez4 device, bluetooth transport was
opened and closed after suspend-on-idle suspended sinks and sources.
This causes problems with PTS tests, so this patch changes
implementation so that bluetooth transport is not opened during device
creation and sink and source are started in suspended state.

[bluez4] Don't acquire transport when creating device. Contributes to JB#27276

Patch 2003-dbus-Use-hooks-for-all-events.patch:

Fixes module-dbus-protocl implementation so that hooks are used instead
of asynchronous subscription system.

[dbus-protocol] Use hooks for all events. Fixes JB#27166