sfos: bluez5: Wait for NewConnection even if Connect fails.
If the initial Connect call fails it is possible that the
headset in question is still starting up and doesn't want
connection from our end just yet. In that case say that we
are connecting and wait for possible NewConnection from
the starting device. In case there isn't going to be any
NewConnection set up a timer to update the backend state.

Signed-off-by: Juho Hämäläinen <juho.hamalainen@jolla.com>