sfos: packaging: Add configuration for system-wide mode
Add configuration for system-wide mode. These new unit is not meant to
be started by user but instead required when making emergency calls
outside user session. PartOf ensures that the unit is stopped when
emergency-call.target unit is stopped so that it doesn't collide with
user session pulseaudio.

There are no conflicts to multi-user.target or
systemd-user-sessions.service because otherwise they may not be started
at all if this unit happens to be still running.